refactor: 카테고리 모달 리팩토링
